Move Store_token function to store_token.c
